refactor(react): update keyboard helper usage
