+++ packages/react/src/components/ToggleSmall/ToggleSmall.js
@@ -9,7 +9,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import classNames from 'classnames';
 import { settings } from 'carbon-components';
-import { keys, match } from '../../tools/key';
+import { keys, match } from '../../internal/keyboard';
 
 const { prefix } = settings;
 
@@ -56,7 +56,7 @@ const ToggleSmall = ({
           input = el;
         }}
         onKeyUp={evt => {
-          if (match(evt, keys.ENTER)) {
+          if (match(evt, keys.Enter)) {
             input.checked = !input.checked;
             onChange(evt);
             onToggle(input.checked, id, evt);
