+++ packages/react/src/components/UIShell/HeaderMenu.js
@@ -10,7 +10,7 @@ import { settings } from 'carbon-components';
 import cx from 'classnames';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { keys, matches } from '../../tools/key';
+import { keys, matches } from '../../internal/keyboard';
 import { AriaLabelPropType } from '../../prop-types/AriaPropTypes';
 
 const { prefix } = settings;
@@ -83,7 +83,7 @@ class HeaderMenu extends React.Component {
    */
   handleOnKeyDown = event => {
     // Handle enter or space key for toggling the expanded state of the menu.
-    if (matches(event, [keys.ENTER, keys.SPACE])) {
+    if (matches(event, [keys.Enter, keys.Space])) {
       event.stopPropagation();
       event.preventDefault();
 
@@ -134,7 +134,7 @@ class HeaderMenu extends React.Component {
 
   handleMenuClose = event => {
     // Handle ESC keydown for closing the expanded menu.
-    if (matches(event, [keys.ESC]) && this.state.expanded) {
+    if (matches(event, [keys.Escape]) && this.state.expanded) {
       event.stopPropagation();
       event.preventDefault();
